requirements.txt
# Trondheim Historical Map - Complete Requirements
# Install with: pip install -r requirements.txt

# ============================================================
# Core ML Dependencies
# ============================================================
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.1

# ============================================================
# Image Processing
# ============================================================
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
scipy>=1.11.0
scikit-image>=0.21.0

# ============================================================
# Geospatial Processing
# ============================================================
shapely>=2.0.2
geojson>=3.1.0
rasterio>=1.3.9
pyproj>=3.6.0

# GDAL - Platform-specific installation required:
#   macOS:   brew install gdal && pip install gdal==$(gdal-config --version)
#   Ubuntu:  apt-get install gdal-bin libgdal-dev && pip install gdal==$(gdal-config --version)
#   Windows: Download from https://www.lfd.uci.edu/~gohlke/pythonlibs/#gdal
# GDAL>=3.6.0

# ============================================================
# Vector Tiles
# ============================================================
pmtiles>=3.3.0
mapbox-vector-tile>=2.0.1

# ============================================================
# Web Backend
# ============================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# ============================================================
# Utilities
# ============================================================
requests>=2.31.0
tqdm>=4.65.0
PyYAML>=6.0.1
python-dateutil>=2.8.0

# ============================================================
# Metrics and Visualization
# ============================================================
scikit-learn>=1.3.0
matplotlib>=3.7.0

# ============================================================
# Optional: Logging and Monitoring
# ============================================================
# tensorboard>=2.14.0
# wandb>=0.16.0

# ============================================================
# Optional: Development Tools
# ============================================================
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.7.0
# flake8>=6.1.0
# mypy>=1.5.0
